feat: move api to api folder
