+++ src/api/variables.ts
@@ -1,0 +1,4 @@
+export const BASE_API_URL = 'https://swapi.dev/api';
+export const PEOPLE_API_URL = `${BASE_API_URL}/people`;
+export const IMAGE_BASE_URL =
+  'https://starwars-visualguide.com/assets/img/characters';
